Refactor: Extract command modules from AMI::Client (Issue #47)
